--- app/src/main/java/com/jaknaeso/app/presentation/Route.kt
@@ -1,5 +1,0 @@
-package com.jaknaeso.app.presentation
-
-enum class Route {
-    Home, MyPage, Login
-}
