+++ app/src/main/java/com/jaknaeso/app/presentation/navigation/NavDestination.kt
@@ -1,0 +1,22 @@
+package com.jaknaeso.app.presentation.navigation
+
+import androidx.navigation.NavController
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.compose.composable
+import com.jaknaeso.app.presentation.home.view.HomeScreen
+import com.jaknaeso.app.presentation.login.LoginScreen
+
+fun NavController.navigateToLogin() = navigate("${Route.Login}")
+fun NavController.navigateToHome() = navigate("${Route.Home}")
+
+fun NavGraphBuilder.loginScreen(navigateToHome: () -> Unit) {
+    composable(route = "${Route.Login}") {
+        LoginScreen(navigateToHome = navigateToHome)
+    }
+}
+
+fun NavGraphBuilder.homeScreen() {
+    composable(route = "${Route.Home}") {
+        HomeScreen()
+    }
+}
