+++ app/src/main/java/com/jaknaeso/app/presentation/navigation/NavHost.kt
@@ -1,13 +1,8 @@
-package com.jaknaeso.app.presentation
+package com.jaknaeso.app.presentation.navigation
 
 import androidx.compose.runtime.Composable
-import androidx.navigation.NavController
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
-import com.jaknaeso.app.presentation.login.loginScreen
-
-fun NavController.navigateToHome() = navigate("${Route.Home}")
-fun NavController.navigateToLogin() = navigate("${Route.Login}")
 
 @Composable
 fun SetUpNavGraph(
@@ -18,6 +13,7 @@ fun SetUpNavGraph(
         navController = navController,
         startDestination = startDestination
     ) {
-        loginScreen(navigateToHome = { navController.navigateToHome() })
+        loginScreen(navigateToHome = navController::navigateToHome)
+        homeScreen()
     }
 }
