+++ app/src/main/java/com/jaknaeso/app/presentation/navigation/Route.kt
@@ -1,0 +1,5 @@
+package com.jaknaeso.app.presentation.navigation
+
+enum class Route {
+    Home, MyPage, Login
+}
